Fix: user middleware fixing and login / logout button state confirm
