+++ views/index.ejs
@@ -34,5 +34,6 @@
     </div>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.7.1/gsap.min.js"></script>
     <script src="/js/index.js"></script>
+    <script src="/js/common.js"></script>
 </body>
 </html>
